Fix float equality comparisons in AmountMatcherRule and TransferMatchingService
AmountMatcherRule used == to compare float values, which fails when two
floats representing the same amount have different bit patterns due to
how they were computed. Replaced with Float.compare() == 0.

TransferMatchingService used Double.compare() on float values, which
promotes floats to double before comparing. While this happened to work
in practice (IEEE 754 negation is exact), using Float.compare() is
clearer and avoids implicit widening.

+++ src/main/java/ca/jonathanfritz/ofxcat/service/TransferMatchingService.java
@@ -44,7 +44,7 @@ public class TransferMatchingService {
         for (Transaction source : sourceTransactions) {
             final List<Transaction> potentialSinks = sinkTransactions.stream()
                     .filter(t -> t.getDate().equals(source.getDate()))
-                    .filter(t -> Double.compare(t.getAmount(), -source.getAmount()) == 0)
+                    .filter(t -> Float.compare(t.getAmount(), -source.getAmount()) == 0)
                     .filter(t -> !t.getAccount().equals(source.getAccount()))
                     .toList();
